datafrag-manager/pyproject.toml
[tool.poetry]
name = "datafrag_manager"
version = "0.0.5"
description = "Data Fragments Manager in dataflows spark"
authors = ["Gustavo Decarlo <gustavodecarlo@gmail.com>"]

[tool.poetry.dependencies]
pyspark = "3.3.2"
cassandra-driver = "^3.25.0"
"boolean.py" = "^4.0"
pyarrow = "^10.0.1"
sqloxide = "^0.1.26"
python = "^3.9"

[tool.poetry.dev-dependencies]
pylint = "^2.8.2"
pytest = "^6.2.4"
mypy = "^0.910"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"